+++ packages/core/src/DropDown.svelte
@@ -1,48 +1,63 @@
 <script>
-	import ListItems from './ListItems.svelte'
+	import Icon from './Icon.svelte'
 	import Text from './items/Text.svelte'
-	import Slider from './Slider.svelte'
+	import ListItems from './ListItems.svelte'
+	import { defaultFields } from './constants.js'
+	import { dismissable } from './actions'
+	import { createEventDispatcher } from 'svelte'
 
-	export let data
-	export let component = null
-	export let selected = null
-	export let key = 'id'
-	export let selectedKey
+	const dispatch = createEventDispatcher()
 
-	let opened = false
+	let className = ''
+	export { className as class }
+	export let items = []
+	export let fields = defaultFields
+	export let using = { default: Text }
+	export let value = null
+	export let placeholder = 'Select a value'
+
+	$: using = { default: Text, ...using }
+	$: fields = { ...defaultFields, ...fields }
+
+	let offsetTop = 0
+	let open = false
 
 	function handleSelect(event) {
-		console.log('selected', event)
-		selected = event.detail
+		open = false
+		dispatch('change', event.detail)
 	}
 </script>
 
 <!-- svelte-ignore a11y-no-noninteractive-tabindex -->
 <div
-	class="flex flex-col outline-none w-full h-12 relative z-1 dropdown"
-	class:opened
-	on:focus={() => (opened = true)}
-	on:blur={() => (opened = false)}
-	tabindex={0}
+	class="flex w-full relative cursor-pointer dropdown {className}"
+	class:open
+	tabindex="0"
+	use:dismissable
+	on:blur={() => (open = false)}
+	on:dismiss={() => (open = false)}
 >
-	<div class="flex flex-shrink-0 h-12 items-center pl-4 selected-item">
-		<span class="flex flex-grow">
-			<Text {component} item={selected} />
-		</span>
-		<icon class="selector" />
-		<!-- <DropDownIcon /> -->
-	</div>
-
-	{#if opened}
-		<Slider class="h-60 absolute top-12 w-full z-1">
+	<button
+		on:click|stopPropagation={() => (open = !open)}
+		class="flex w-full"
+		bind:clientHeight={offsetTop}
+		tabindex="-1"
+	>
+		<svelte:component this={using.default} content={value ?? placeholder} />
+		<Icon name="dropdown-opened" />
+	</button>
+	{#if open}
+		<div
+			class="flex flex-col absolute z-10 h-fit w-full menu"
+			style:top="{offsetTop}px"
+		>
 			<ListItems
-				bind:items={data}
-				{key}
-				{component}
-				bind:selected={selectedKey}
+				{items}
+				{fields}
+				{using}
+				bind:activeItem={value}
 				on:select={handleSelect}
-				on:change
 			/>
-		</Slider>
+		</div>
 	{/if}
 </div>