+++ packages/layout/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rokkit/layout
 
+## 1.0.0-next.9
+
+### Patch Changes
+
+- Updated dropdown and theme
+- Updated dependencies
+  - @rokkit/core@1.0.0-next.9
+
 ## 1.0.0-next.5
 
 ### Patch Changes